Ничего особенного
Функционал создания привычки еще не создан. Тут просто работа над ошибками
